# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Feelms streaming platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /interpreters:
    get:
      operationId: getInterpreters
      tags: [interpreters]
      summary: List all interpreters
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interpreter"
    post:
      operationId: createInterpreter
      tags: [interpreters]
      summary: Create an interpreter (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterpreterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interpreter"
  /interpreters/{id}:
    patch:
      operationId: updateInterpreter
      tags: [interpreters]
      summary: Update an interpreter (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterpreterUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interpreter"
    delete:
      operationId: deleteInterpreter
      tags: [interpreters]
      summary: Delete an interpreter (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # BANNERS
...
  /sections:
    get:
      operationId: getSections
      tags: [sections]
      summary: List homepage sections
      parameters:
        - name: all
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"
    post:
      operationId: createSection
      tags: [sections]
      summary: Create a section (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
  /sections/{id}:
    patch:
      operationId: updateSection
      tags: [sections]
      summary: Update a section (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
    delete:
      operationId: deleteSection
      tags: [sections]
      summary: Delete a section (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # USERS
...
  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # PAYMENTS
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate a VIP subscription payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitResponse"
  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify a payment and upgrade user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerifyInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResponse"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [FREE, VIP, ADMIN]
        vipExpiry:
          type: ["string", "null"]
        createdAt:
          type: string
    RoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [FREE, VIP, ADMIN]
        vipExpiry:
          type: ["string", "null"]

    # INTERPRETERS
    Interpreter:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        bio:
          type: ["string", "null"]
        photo:
          type: ["string", "null"]
        createdAt:
          type: string
    InterpreterInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        bio:
          type: string
        photo:
          type: string
    InterpreterUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        photo:
          type: string

    # MOVIES
    Movie:
      type: object
      required: [id, title, description, poster, genres, vipOnly, featured, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        poster:
          type: string
        trailer:
          type: ["string", "null"]
        streamingLinks:
          type: array
          items:
            type: string
        downloadLinks:
          type: array
          items:
            type: string
        vipOnly:
          type: boolean
        featured:
          type: boolean
        genres:
          type: array
          items:
            type: string
        year:
          type: ["integer", "null"]
        rating:
          type: ["number", "null"]
        duration:
          type: ["string", "null"]
        watchCount:
          type: integer
        interpreted:
          type: boolean
        interpreters:
          type: array
          items:
            type: string
        createdAt:
          type: string
...
    Section:
      type: object
      required: [id, title, movieIds, orderIndex, enabled, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        movieIds:
          type: array
          items:
            type: integer
        orderIndex:
          type: integer
        enabled:
          type: boolean
        sectionType:
          type: ["string", "null"]
        createdAt:
          type: string
        movies:
          type: array
          items:
            $ref: "#/components/schemas/Movie"
    SectionInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        movieIds:
          type: array
          items:
            type: integer
        orderIndex:
          type: integer
        enabled:
          type: boolean
        sectionType:
          type: string
    SectionUpdate:
      type: object
      properties:
        title:
          type: string
        movieIds:
          type: array
          items:
            type: integer
        orderIndex:
          type: integer
        enabled:
          type: boolean
        sectionType:
          type: string

    # PAYMENTS
    PaymentInput:
      type: object
      required: [plan, provider, phone]
      properties:
        plan:
          type: string
          enum: [monthly, yearly]
        provider:
          type: string
          enum: [MTN, AIRTEL, FLUTTERWAVE]
        phone:
          type: string
    PaymentInitResponse:
      type: object
      required: [paymentId, status, message]
      properties:
        paymentId:
          type: string
        status:
          type: string
        message:
          type: string
        redirectUrl:
          type: ["string", "null"]
    PaymentVerifyInput:
      type: object
      required: [paymentId]
      properties:
        paymentId:
          type: string
        transactionId:
          type: string
    PaymentVerifyResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        user:
          $ref: "#/components/schemas/User"
...